Updating module and module_util name to amqp.
