+++ test/integration/targets/amqp_publish/aliases
