+++ test/integration/targets/amqp_publish/files/image.gif
