+++ test/integration/targets/amqp_publish/meta/main.yml
