+++ test/integration/targets/amqp_publish/tasks/main.yml
