+++ test/integration/targets/amqp_publish/tasks/ubuntu.yml
@@ -3,66 +3,66 @@
     name: requests,pika
     state: present
 
-- name: RabbitMQ basic publish test
-  rabbitmq_publish:
+- name: AMQP basic publish test
+  amqp_publish:
     url: "amqp://guest:guest@localhost:5672/%2F"
     queue: 'publish_test'
-    body: "Hello world from ansible module rabbitmq_publish"
+    body: "Hello world from ansible module amqp_publish"
     content_type: "text/plain"
-  register: rabbit_basic_output1
+  register: amqp_basic_output1
 
 - assert:
     that:
-      - "rabbit_basic_output1 is not failed"
-      - "'publish_test' in rabbit_basic_output1.result.msg"
-      - "'publish_test' in rabbit_basic_output1.result.queue"
-      - "'text/plain' in rabbit_basic_output1.result.content_type"
+      - "amqp_basic_output1 is not failed"
+      - "'publish_test' in amqp_basic_output1.result.msg"
+      - "'publish_test' in amqp_basic_output1.result.queue"
+      - "'text/plain' in amqp_basic_output1.result.content_type"
 
 
 # Testing random queue
 - name: Publish to random queue
-  rabbitmq_publish:
+  amqp_publish:
     url: "amqp://guest:guest@localhost:5672/%2F"
     body: "RANDOM QUEUE POST"
     content_type: "text/plain"
-  register: rabbit_random_queue_output
+  register: amqp_random_queue_output
 
 - assert:
     that:
-      - "rabbit_random_queue_output is not failed"
-      - "'amq.gen' in rabbit_random_queue_output.result.msg"
-      - "'amq.gen' in rabbit_random_queue_output.result.queue"
-      - "'text/plain' in rabbit_random_queue_output.result.content_type"
+      - "amqp_random_queue_output is not failed"
+      - "'amq.gen' in amqp_random_queue_output.result.msg"
+      - "'amq.gen' in amqp_random_queue_output.result.queue"
+      - "'text/plain' in amqp_random_queue_output.result.content_type"
 
 - name: Publish binary to a queue
-  rabbitmq_publish:
+  amqp_publish:
     url: "amqp://guest:guest@localhost:5672/%2F"
     queue: publish_test
     src: "{{ role_path }}/files/image.gif"
-  register: rabbitmq_publish_file
+  register: amqp_publish_file
 
 - assert:
     that:
-      - "rabbitmq_publish_file is not failed"
-      - "'publish_test' in rabbitmq_publish_file.result.queue"
-      - "'image/gif' in rabbitmq_publish_file.result.content_type"
+      - "amqp_publish_file is not failed"
+      - "'publish_test' in amqp_publish_file.result.queue"
+      - "'image/gif' in amqp_publish_file.result.content_type"
 
 - name: Raise error for src and body defined
-  rabbitmq_publish:
+  amqp_publish:
     url: "amqp://guest:guest@localhost:5672/%2F"
     queue: 'publish_test'
     src: "{{ role_path }}/files/image.gif"
     body: blah
-  register: rabbit_basic_fail_output1
+  register: amqp_basic_fail_output1
   ignore_errors: yes
 
 - assert:
     that:
-      - "rabbit_basic_fail_output1 is failed"
-      - "'parameters are mutually exclusive' in rabbit_basic_fail_output1.msg"
+      - "amqp_basic_fail_output1 is failed"
+      - "'parameters are mutually exclusive' in amqp_basic_fail_output1.msg"
 
 - name: Publish a file that does not exist
-  rabbitmq_publish:
+  amqp_publish:
     url: "amqp://guest:guest@localhost:5672/%2F"
     queue: 'publish_test'
     src: 'aaaaaaajax-loader.gif'
@@ -75,7 +75,7 @@
       - "'Unable to open file' in file_missing_fail.msg"
 
 - name: Publish with proto/host/port/user/pass
-  rabbitmq_publish:
+  amqp_publish:
     proto: amqp
     host: localhost
     port: 5672
@@ -91,7 +91,7 @@
       - "host_port_output is not failed"
 
 - name: Publish with host/port/user but missing proto
-  rabbitmq_publish:
+  amqp_publish:
     host: localhost
     port: 5672
     username: guest
@@ -108,7 +108,7 @@
       - "'Connection parameters must be passed via' in host_port_missing_proto_output.msg"
 
 - name: Publish with proto/host/port/user and url
-  rabbitmq_publish:
+  amqp_publish:
     url: "amqp://guest:guest@localhost:5672/%2F"
     proto: amqp
     host: localhost
@@ -127,7 +127,7 @@
       - "'cannot be specified at the same time' in host_and_url_output.msg"
 
 - name: Publish headers to queue
-  rabbitmq_publish:
+  amqp_publish:
     url: "amqp://guest:guest@localhost:5672/%2F"
     queue: 'publish_test'
     body: blah
@@ -138,7 +138,7 @@
   ignore_errors: yes
 
 - name: Publish headers with file
-  rabbitmq_publish:
+  amqp_publish:
     url: "amqp://guest:guest@localhost:5672/%2F"
     queue: 'publish_test'
     src: "{{ role_path }}/files/image.gif"
@@ -156,7 +156,7 @@
   assert:
     that:
       - messages|length == 5
-      - "'Hello world from ansible module rabbitmq_publish' in messages[0]['msg']"
+      - "'Hello world from ansible module amqp_publish' in messages[0]['msg']"
       - "'text/plain' in messages[0]['content_type']"
       - "'image/gif' in messages[1]['content_type']"
       - "'image.gif' in messages[1]['headers']['filename']"
